requirements.txt
aiogram==3.26.0
python-dotenv
yt-dlp
telethon
